Start to use a Sqlite database to store mapping data
I've been thinking for a while about using a SQL database to store the
MCP mapping information and possibly even cache already computed.
One night it occured to me that if I stored everything in SQL,
it would not only simplify handling my data it would allow me to use
joins to quickly chain mappings together.

I'm already quite pleased with the cleanliness of this solution.
It was realatively easy and quick to load the obfuscated mappings data
into SQLite.
SQL transactions provide a much better alternative to my
hand-written file-based cache, and automatically rollback to the last
state even if I panic or accidently shut off the database.
JOINS should provide a much faster alternative to the old mappings
system, without having to allocate and hash large numbers of strings.

Although we still use the cache from the `minecraft-mappings` crate,
it should be rarely used since we really only use it to load the
original mapping data.

Once I get the new database up and running,
I'll finish writing the database CLI and start to collect timings.
If it's still not as fast as I think it should be,
I'll probably have to start optimizing my SQL :vomit: