First attempt at a SQL mappings database
Even the easiest database query can't is three times slower than the
hardest target in the old system.
This database design is unacceptably slow and I don't
want to invest further effort into it.

I actually do need a database eventually,
since most of the queries to
the REST API will probably be for single methods and fields
and not for the mappings as a whole.